Shortcut asprintf_impl if length < 1
If the output string length will be less than 1, then there is no
point going through the motions of allocating a buffer, copying an
empty string, etc. Since -1 indicates an error, this only really has
an effect when the length would be 0 as unistd::asprintf_impl was
already returning when length < 0.